% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{boot1}
\alias{boot1}
\title{Boots}
\format{511x198 matrices representing scans from two different
    boots.}
\usage{
boot1
}
\description{
Sample scans of boots
}
\details{
These images have a very low resolution (compared to the images
that this packages is intended to handle) due to the reason that
they will be included in the package.
}
\examples{
dim(boot1)
dim(boot2)
image(boot1)
}
\keyword{datasets}
