% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_moments.R
\name{calculate_central_moments}
\alias{calculate_central_moments}
\title{Normalized central moments of an image}
\usage{
calculate_central_moments(img, p, q, x_centered, y_centered)
}
\arguments{
\item{img}{An image stored in a matrix.}

\item{p}{x-moment.}

\item{q}{y-moment.}

\item{x_centered}{Vector of centered coordinates.}

\item{y_centered}{Vector of centered coordinates.}
}
\value{
A real value with the qp normalized moment.
}
\description{
This function provides the normalized central moments of an image.
}
\author{
Guillermo Basulto-Elias
}
