% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_moments.R
\name{compute_centroid}
\alias{compute_centroid}
\title{Centroid of an image}
\usage{
compute_centroid(img, x_grid, y_grid)
}
\arguments{
\item{img}{An image represented by a matrix}

\item{x_grid}{Vector with the coordinates of x}

\item{y_grid}{Vector with coordinates of y}
}
\value{
A vector of size two with the centroid.
}
\description{
Computes the centroid of an image.
}
\author{
Guillermo Basulto-Elias
}
