% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_moments.R
\name{compute_moments}
\alias{compute_moments}
\title{Hu moments}
\usage{
compute_moments(img)
}
\arguments{
\item{img}{A matrix representing and image}
}
\value{
A numeric vector of size 7 with the moments.
}
\description{
This function computes Hu's moments from an image
}
\author{
Guillermo Basulto-Elias
}
