% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_images.R
\name{register_images}
\alias{register_images}
\title{Register Images}
\usage{
register_images(img_orig, img_trans, img_orig_p1, img_orig_p2, img_orig_p3,
  img_trans_p1, img_trans_p2, img_trans_p3)
}
\arguments{
\item{img_orig}{A matrix representing the original image}

\item{img_trans}{A matrix representing the translated image}

\item{img_orig_p1}{Reference point 1 in the original image}

\item{img_orig_p2}{Reference point 2 in the original image}

\item{img_orig_p3}{Reference point 3 in the original image}

\item{img_trans_p1}{Reference point 1 in the translated image}

\item{img_trans_p2}{Reference point 2 in the translated image}

\item{img_trans_p3}{Reference point 3 in the translated image}
}
\value{
The registration of the two images
}
\description{
This function register two images based on two points
selected by the user. This function is useful in shiny app.
}
\author{
Martin Silerio-Vazquez
}
