% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_image.R
\name{resize_image}
\alias{resize_image}
\title{Resize Image}
\usage{
resize_image(img, fraction)
}
\arguments{
\item{img}{A image passed as a matrix.}

\item{fraction}{Proportion of memory of the resulting image.}
}
\value{
An image
}
\description{
Resize an image to a fraction of its space.
}
\author{
Guillermo Basulto-Elias
}
