% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{update_click}
\alias{update_click}
\title{Update click}
\usage{
update_click(ranges, click)
}
\arguments{
\item{ranges}{A list with the limits in x and y where the image
was zoomed-in.}

\item{click}{Point where click was done in the scale of the
zoomed-in image.}
}
\value{
A dataframe with x and y in the original scale.
}
\description{
This function is used in the shiny app. A large image is zoomed-in
and a click is done on the second image. This function matches the
click with the coordinates of the original image.
}
\author{
Guillermo Basulto-Elias
}
